refactor(llm): reorganize module structure for better separation of concerns
Extracted model profiles, parameter handling, and factory logic from the monolithic client.py into separate modules:
- Created model_profiles.py for MODEL_PATTERNS and PARAMETER_PROFILES
- Created parameter_handler.py for parameter conversion and model family functions
- Created factory.py for LLMClientFactory
- Updated imports across __init__.py, api_handlers.py, and client.py

This refactoring reduces client.py from 793+ lines to a more focused module, improves code maintainability, and follows the single responsibility principle. No functional changes were made.
