+++ litassist/llm/__init__.py
@@ -14,18 +14,17 @@ from .api_handlers import (
     NonRetryableAPIError,
 )
 
-from .client import (
-    LLMClient,
-    LLMClientFactory,
+from .model_profiles import MODEL_PATTERNS, PARAMETER_PROFILES
+from .parameter_handler import (
+    convert_thinking_effort,
+    convert_verbosity,
     get_model_family,
     get_model_parameters,
     get_openrouter_params,
     supports_system_messages,
-    convert_thinking_effort,
-    convert_verbosity,
-    MODEL_PATTERNS,
-    PARAMETER_PROFILES,
 )
+from .factory import LLMClientFactory
+from .client import LLMClient
 
 from .verification import (
     LLMVerificationMixin,
