+++ litassist/llm/client.py
@@ -5,793 +5,37 @@ This module provides a unified interface for chat completions across different L
 handling parameter management and response processing.
 """
 
-import re
-import os
+import time
+import logging
 from datetime import datetime
 from typing import List, Dict, Any, Tuple
 
 from litassist.timing import timed
 from litassist.logging_utils import save_log
 from litassist.utils.core import heartbeat
-from litassist.utils.formatting import info_message, success_message
+from litassist.utils.formatting import success_message
 from litassist.config import get_config
 from litassist.prompts import PROMPTS
-import time
-from litassist.citation_verify import (
-    CitationVerificationError,
-)
+from litassist.citation_verify import CitationVerificationError
+
 from .api_handlers import execute_api_call_with_retry
 from .verification import LLMVerificationMixin
 from .response_parser import extract_content_and_usage
 from .retry_handler import handle_citation_retry
 from .citation_handler import process_citation_verification, handle_retry_failure
 from .tools import get_tool_definitions, execute_tool, format_tool_response
-
-import logging
+from .parameter_handler import (
+    get_model_family,
+    get_model_parameters,
+    supports_system_messages,
+)
 
 
 logger = logging.getLogger(__name__)
 
 
-# Model family patterns for dynamic parameter handling
-MODEL_PATTERNS = {
-    "openai_reasoning": r"openai/o\d+",  # Matches o1, o3, o1-pro, o3-pro, o4, etc.
-    "gpt5": r"openai/gpt-5(-pro)?",  # GPT-5 and GPT-5 Pro (August/October 2025)
-    "claude4": r"anthropic/claude-(opus-4|sonnet-4)(\.\d+)?",  # Claude 4 models (includes 4.1, 4.5, etc.)
-    "anthropic": r"anthropic/claude",  # Other Claude models
-    "google": r"google/(gemini|palm|bard)",
-    "openai_standard": r"openai/(gpt|chatgpt)",  # GPT-4, ChatGPT, etc.
-    "xai": r"x-ai/grok",
-    "meta": r"meta/(llama|codellama)",
-    "mistral": r"mistral/",
-    "cohere": r"cohere/",
-    "moonshotai": r"moonshotai/",
-}
-
-# Parameter profiles by model family
-PARAMETER_PROFILES = {
-    "openai_reasoning": {
-        "allowed": [
-            "max_completion_tokens",
-            "reasoning",  # OpenRouter reasoning object
-            "verbosity",  # GPT-5 and newer models
-            "seed",
-            "response_format",
-            "structured_outputs",
-        ],
-        "transforms": {
-            "max_tokens": "max_completion_tokens",
-        },
-        "system_message_support": False,  # o1/o3 models don't support system messages (but DO support tools as of 2025)
-    },
-    "anthropic": {
-        "allowed": [
-            "temperature",
-            "top_p",
-            "max_tokens",
-            "stop",
-            "top_k",
-            "stream",
-            "metadata",
-            "stop_sequences",
-            "reasoning",  # OpenRouter reasoning object
-            # Advanced parameters
-            "min_p",
-            "top_a",
-            "repetition_penalty",
-        ],
-        "transforms": {},
-    },
-    "google": {
-        "allowed": [
-            "temperature",
-            "top_p",
-            "max_tokens",
-            "stop",
-            "candidate_count",
-            "top_k",
-            "safety_settings",
-            "stop_sequences",
-            "reasoning",  # OpenRouter reasoning object (if supported)
-            # Advanced parameters
-            "min_p",
-            "top_a",
-        ],
-        "transforms": {},
-    },
-    "openai_standard": {
-        "allowed": [
-            "temperature",
-            "top_p",
-            "max_tokens",
-            "frequency_penalty",
-            "presence_penalty",
-            "stop",
-            "logit_bias",
-            "seed",
-            "response_format",
-            "stream",
-            "n",
-            "tools",
-            "tool_choice",
-            "functions",
-            "function_call",
-            "user",
-            "logprobs",
-            "top_logprobs",
-        ],
-    },
-    "xai": {
-        # OpenRouter-specific parameters (min_p, top_a, repetition_penalty) are handled
-        # through extra_body in api_handlers.py, not as direct parameters
-        "allowed": [
-            "temperature",
-            "top_p",
-            "max_tokens",
-            "stop",
-            "frequency_penalty",
-            "presence_penalty",
-            "stream",
-            "reasoning",  # Grok models support reasoning
-            "verbosity",
-            # OpenRouter-specific params removed from here, handled via extra_body
-        ],
-    },
-    "meta": {
-        "allowed": [
-            "temperature",
-            "top_p",
-            "max_tokens",
-            "stop",
-            "frequency_penalty",
-            "presence_penalty",
-            "stream",
-            "min_p",
-            "top_a",
-            "repetition_penalty",
-        ],
-    },
-    "mistral": {
-        "allowed": [
-            "temperature",
-            "top_p",
-            "max_tokens",
-            "stop",
-            "random_seed",
-            "safe_mode",
-            "stream",
-            "min_p",
-            "top_a",
-            "repetition_penalty",
-        ],
-        "transforms": {"seed": "random_seed"},
-    },
-    "cohere": {
-        "allowed": [
-            "temperature",
-            "max_tokens",
-            "k",
-            "p",
-            "stop_sequences",
-            "frequency_penalty",
-            "presence_penalty",
-            "stream",
-        ],
-        "transforms": {"top_k": "k", "top_p": "p", "stop": "stop_sequences"},
-    },
-    "moonshotai": {
-        "allowed": [
-            "temperature",
-            "top_p",
-            "max_tokens",
-            "stop",
-            "frequency_penalty",
-            "presence_penalty",
-            "stream",
-            "min_p",
-            "top_a",
-            "repetition_penalty",
-        ],
-    },
-    "default": {
-        "allowed": ["temperature", "top_p", "max_tokens", "stop"],  # Safe defaults
-    },
-    # Universal parameters supported by OpenRouter across models
-    "openrouter_universal": {
-        "allowed": [
-            "temperature",
-            "top_p",
-            "max_tokens",
-            "reasoning",  # OpenRouter unified reasoning object
-            "verbosity",
-            "min_p",
-            "top_a",
-            "repetition_penalty",
-            "frequency_penalty",
-            "presence_penalty",
-            "response_format",
-            "logit_bias",
-            "logprobs",
-            "top_logprobs",
-            "seed",
-            "stop",
-        ],
-    },
-}
-
-
-def convert_thinking_effort(effort: str, model_name: str) -> dict:
-    """
-    Convert universal thinking_effort to OpenRouter's reasoning object format.
-
-    Args:
-        effort: Universal effort level (none, minimal, low, medium, high, max)
-        model_name: Full model name (e.g., "openai/o3-pro", "anthropic/claude-4")
-
-    Returns:
-        Dict with OpenRouter reasoning object
-    """
-
-    if effort == "none":
-        return {}  # Don't send reasoning parameter
-
-    # OpenRouter unified reasoning object approach - ALL models go through OpenRouter
-    model_family = get_model_family(model_name)
-
-    # Check model type for appropriate sub-parameters
-    if model_family in ["openai_reasoning", "gpt5", "xai"]:
-        # Effort-based models (OpenAI, Grok, GPT-5)
-        effort_map = {
-            "minimal": "minimal",  # GPT-5 specific
-            "low": "low",
-            "medium": "medium",
-            "high": "high",
-            "max": "high",  # Map max to highest available
-        }
-        mapped_effort = effort_map.get(effort, "medium")
-
-        # Only include minimal for GPT-5 and o4-mini
-        if (
-            mapped_effort == "minimal"
-            and model_family not in ["gpt5"]
-            and "o4" not in model_name
-        ):
-            mapped_effort = "low"  # Fallback for non-GPT-5/o4 models
-
-        # Special handling for o4-mini with summary field
-        if "o4" in model_name:
-            return {
-                "reasoning": {
-                    "effort": mapped_effort,
-                    "summary": "auto",  # New o4 feature for automatic summarization
-                }
-            }
-        # GPT-5 supports both reasoning and verbosity
-        elif model_family == "gpt5":
-            return {
-                "reasoning": {"effort": mapped_effort}
-                # Verbosity handled separately via convert_verbosity
-            }
-        else:
-            return {"reasoning": {"effort": mapped_effort}}
-
-    elif model_family in ["claude4", "anthropic"]:
-        # Token-based models (Anthropic)
-        token_map = {
-            "minimal": 1024,
-            "low": 1024,
-            "medium": 8192,
-            "high": 16384,
-            "max": 32000,  # Max allowed by OpenRouter
-        }
-        return {"reasoning": {"max_tokens": token_map.get(effort, 8192)}}
-
-    elif model_family == "google":
-        # Google/Gemini models - try unified reasoning
-        effort_map = {
-            "minimal": "low",
-            "low": "low",
-            "medium": "medium",
-            "high": "high",
-            "max": "high",
-        }
-        return {"reasoning": {"effort": effort_map.get(effort, "medium")}}
-
-    # For all other models, don't add reasoning parameters
-    return {}
-
-
-def convert_verbosity(level: str, model_name: str = None) -> dict:
-    """
-    Convert verbosity level to API parameter.
-
-    Args:
-        level: Verbosity level (low, medium, high)
-        model_name: Optional model name for model-specific handling
-
-    Returns:
-        Dict with verbosity parameter if valid
-    """
-    if level in ["low", "medium", "high"]:
-        # GPT-5 and other models that support verbosity
-        return {"verbosity": level}
-    return {}
-
-
-def get_model_family(model_name: str) -> str:
-    """
-    Identify the model family based on pattern matching.
-
-    Args:
-        model_name: The full model name (e.g., "openai/gpt-4", "anthropic/claude-3")
-
-    Returns:
-        The model family name (e.g., "openai_standard", "anthropic")
-    """
-    for family, pattern in MODEL_PATTERNS.items():
-        if re.match(pattern, model_name):
-            return family
-    return "default"
-
-
-def get_openrouter_params() -> set:
-    """
-    Get the set of OpenRouter-specific parameters that need special handling.
-
-    These parameters are not part of the standard OpenAI API and must be
-    passed through extra_body when using the OpenAI SDK with OpenRouter.
-
-    Returns:
-        Set of parameter names that are OpenRouter-specific
-    """
-    return {"reasoning", "min_p", "top_a", "repetition_penalty"}
-
-
-def get_model_parameters(model_name: str, requested_params: dict) -> dict:
-    """
-    Dynamically filter parameters based on model patterns.
-
-    Returns only the parameters that the model supports,
-    with any necessary transformations applied.
-
-    Args:
-        model_name: The full model name
-        requested_params: Dictionary of requested parameters
-
-    Returns:
-        Filtered dictionary containing only supported parameters
-    """
-    # All models go through OpenRouter
-    model_family = get_model_family(model_name)
-    profile = PARAMETER_PROFILES.get(model_family, PARAMETER_PROFILES["default"])
-
-    filtered = {}
-    transforms = profile.get("transforms", {})
-    allowed = profile.get("allowed", [])
-
-    # Copy parameters to avoid modifying original
-    params_to_process = requested_params.copy()
-
-    # Handle thinking_effort conversion FIRST (highest priority)
-    if (
-        "thinking_effort" in params_to_process
-        and params_to_process["thinking_effort"] is not None
-    ):
-        effort = params_to_process.pop("thinking_effort")
-        reasoning_params = convert_thinking_effort(effort, model_name)
-        filtered.update(reasoning_params)
-
-        # CRITICAL: Remove any conflicting parameters to prevent API errors
-        # OpenRouter doesn't allow both 'reasoning' and 'reasoning_effort'
-        params_to_process.pop("reasoning_effort", None)
-        params_to_process.pop("reasoning", None)
-        params_to_process.pop("thinking", None)
-        params_to_process.pop("thinking_config", None)
-
-    # Handle verbosity parameter
-    if "verbosity" in params_to_process and params_to_process["verbosity"] is not None:
-        verbosity = params_to_process.pop("verbosity")
-        verbosity_params = convert_verbosity(verbosity, model_name)
-        filtered.update(verbosity_params)
-
-    # Get OpenRouter-specific parameters
-    openrouter_params = get_openrouter_params()
-
-    # Process remaining parameters
-    for param, value in params_to_process.items():
-        # Skip None values
-        if value is None:
-            continue
-
-        # Check if parameter needs transformation
-        if param in transforms:
-            new_param = transforms[param]
-            filtered[new_param] = value
-        elif param in allowed:
-            filtered[param] = value
-        elif param in openrouter_params:
-            # Preserve OpenRouter-specific params - they'll be moved to extra_body in api_handlers
-            filtered[param] = value
-        # Silently drop other unsupported parameters
-        # Note: We don't add universal parameters automatically to maintain model-specific restrictions
-
-    return filtered
-
-
-def supports_system_messages(model_name: str) -> bool:
-    """
-    Check if a model supports system messages.
-
-    Args:
-        model_name: The full model name
-
-    Returns:
-        True if the model supports system messages, False otherwise
-    """
-    model_family = get_model_family(model_name)
-    profile = PARAMETER_PROFILES.get(model_family, PARAMETER_PROFILES["default"])
-    return profile.get("system_message_support", True)  # Default to True
-
-
-class LLMClientFactory:
-    """
-    Factory class for creating LLMClient instances with command-specific configurations.
-
-    All models use "provider/model" format and route through OpenRouter.
-
-    This centralizes all model and parameter configurations for each command,
-    eliminating duplication and providing a single source of truth.
-    """
-
-    # Command configurations registry
-    COMMAND_CONFIGS = {
-        # Extract facts - deterministic, focused on accuracy
-        # UPGRADED: Oct 2025 - Sonnet 4 -> Sonnet 4.5 for better legal domain knowledge
-        "extractfacts": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-sonnet-4",
-            "temperature": 0,
-            "top_p": 0.15,
-            "thinking_effort": "high",  # Critical foundational command needs thorough thinking
-            "enforce_citations": True,  # Retry on citation errors for foundational docs
-            "disable_tools": True,  # Check if Sonnet 4.5 still has tool calling issues
-        },
-        # Strategy - enhanced multi-step legal reasoning
-        # UPGRADED: Oct 2025 - Opus 4.1 -> Sonnet 4.5 (state-of-the-art for litigation, 80% cost reduction)
-        "strategy": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-opus-4.1",
-            "temperature": 0.2,  # Controlled creativity for strategic thinking
-            "top_p": 0.8,  # Focused but not overly restrictive
-            "thinking_effort": "max",  # Universal parameter, translates to reasoning object
-            "verbosity": "medium",  # Balanced depth in strategic analysis
-            "enforce_citations": False,  # Don't retry on citation errors for strategy
-        },
-        # Strategy sub-type for analysis
-        "strategy-analysis": {
-            "model": "openai/o3-pro",
-            # Note: o3-pro ignores temperature and top_p parameters
-            "thinking_effort": "max",  # Universal parameter, translates to reasoning_effort
-            "disable_tools": True,  # o3-pro doesn't support tool calling
-        },
-        # Brainstorm - varied temperatures for different approaches
-        # UPGRADED: Oct 2025 - Opus 4.1 -> Sonnet 4.5 (better legal domain knowledge, cost-effective)
-        "brainstorm-orthodox": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-opus-4.1",
-            "temperature": 0.3,
-            "top_p": 0.7,
-            "thinking_effort": "medium",  # Moderate thinking for balanced analysis
-            "enforce_citations": False,  # Don't retry on citation errors for brainstorm
-        },
-        "brainstorm-unorthodox": {
-            "model": "x-ai/grok-4",
-            "temperature": 0.8,
-            "top_p": 0.95,
-            "min_p": 0.05,  # Allow more diverse token selection
-            "repetition_penalty": 1.2,  # Reduce repetitive ideas
-            # Kimi-K2 currently has an 8K context window. Supplying an
-            # excessively high `max_tokens` causes "Error processing stream".
-            # Explicitly cap it so the request succeeds.
-            # "max_tokens": 4096,
-            "enforce_citations": False,  # Don't retry on citation errors for creative outputs
-        },
-        "brainstorm-analysis": {
-            "model": "openai/o3-pro",
-            "temperature": 0.2,
-            "top_p": 0.8,
-            "thinking_effort": "high",  # Universal parameter, translates to reasoning_effort
-            "disable_tools": True,  # o3-pro doesn't support tool calling
-        },
-        # Draft - superior technical writing (o3 model with very limited parameter support)
-        "draft": {
-            "model": "openai/o3-pro",
-            "thinking_effort": "high",  # Universal parameter
-            "verbosity": "high",  # Comprehensive legal drafting
-            "disable_tools": True,  # o3-pro doesn't support tool calling
-        },
-        # Digest - mode-dependent settings
-        # UPGRADED: Oct 2025 - Sonnet 4 -> Sonnet 4.5 for better reasoning
-        "digest-summary": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-sonnet-4",
-            "temperature": 0.2,
-            "top_p": 0.3,  # Fixed: was 0, too restrictive
-            "thinking_effort": "medium",  # Simple summarization task
-            "disable_tools": True,  # Check if Sonnet 4.5 still has tool calling issues
-        },
-        # UPGRADED: Oct 2025 - Opus 4.1 -> Sonnet 4.5 (state-of-the-art for legal analysis)
-        "digest-issues": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-opus-4.1",
-            "temperature": 0.2,
-            "top_p": 0.5,
-            "thinking_effort": "high",  # Deep analysis for issue spotting
-        },
-        # Lookup - uses Gemini for rapid processing with verification
-        # IMPORTANT: When changing models, adjust max_content_tokens in lookup.py
-        # Current: Gemini 2.5 Pro with 1M context window (using 900k for content)
-        # Other models have smaller limits - see lookup.py line 528
-        "lookup": {
-            "model": "google/gemini-2.5-pro",
-            "temperature": 0.2,
-            "top_p": 0.4,
-            "thinking_effort": "low",  # Fast thinking for rapid search results
-            "verbosity": "low",  # Concise search summaries
-            "enforce_citations": False,  # Don't enforce strict citation retries
-        },
-        # Verification - automatic verification for high-risk commands
-        # UPGRADED: Oct 2025 - Opus 4.1 -> GPT-5 base (1.4% hallucination, fast verification)
-        "verification": {
-            "model": "openai/gpt-5",
-            # OLD: "model": "anthropic/claude-opus-4.1",
-            "temperature": 0.2,
-            "top_p": 0.3,
-            "thinking_effort": "high",
-            "enforce_citations": False,  # Don't double-enforce since this IS verification
-        },
-        # UPGRADED: Oct 2025 - Sonnet 4 -> Sonnet 4.5 for better legal domain knowledge
-        "verification-light": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-sonnet-4",
-            "temperature": 0.2,  # Optimal for factual tasks per hallucination report
-            "top_p": 0.2,  # Focused beam for consistency
-            "thinking_effort": "medium",  # Just spelling/terminology checks
-            "enforce_citations": False,  # Avoid loops
-            "disable_tools": True,  # Check if Sonnet 4.5 still has tool calling issues
-        },
-        # UPGRADED: Oct 2025 - GPT-5 -> GPT-5 Pro (<1% hallucination, 400K context, mandatory reasoning)
-        "verification-heavy": {
-            "model": "openai/gpt-5-pro",
-            # OLD: "model": "openai/gpt-5",
-            "temperature": 0.2,  # Optimal per hallucination report
-            "top_p": 0.3,  # Slightly wider beam for comprehensive checking
-            "thinking_effort": "max",  # Mandatory deep reasoning for critical verification
-            "enforce_citations": False,  # Avoid loops
-        },
-        # Verify sub-commands with specific model assignments
-        # UPGRADED: Oct 2025 - o3-pro -> Sonnet 4.5 (state-of-the-art for legal reasoning extraction)
-        "verify-reasoning": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "openai/o3-pro",
-            "temperature": 0.2,
-            "top_p": 0.3,
-            "thinking_effort": "high",  # Extended thinking for complex reasoning trace extraction
-            "enforce_citations": False,
-            "disable_tools": True,  # Check if Sonnet 4.5 still has tool calling issues
-        },
-        # UPGRADED: Oct 2025 - Opus 4.1 -> GPT-5 Pro (<1% hallucination for critical soundness checking)
-        "verify-soundness": {
-            "model": "openai/gpt-5-pro",
-            # OLD: "model": "anthropic/claude-opus-4.1",
-            "temperature": 0.2,
-            "top_p": 0.3,
-            "thinking_effort": "max",  # Mandatory deep reasoning for soundness analysis
-            "verbosity": "high",  # Detailed soundness explanations
-            "enforce_citations": False,
-        },
-        # Counsel's Notes - strategic analysis from advocate's perspective
-        "counselnotes": {
-            "model": "openai/o3-pro",
-            "temperature": 0.3,
-            "top_p": 0.7,
-            "thinking_effort": "high",  # Universal parameter, translates to reasoning_effort
-            "enforce_citations": False,  # Don't retry on citation errors for counsel's notes
-            "disable_tools": True,  # o3-pro doesn't support tool calling
-        },
-        # Barrister's brief - comprehensive document generation
-        "barbrief": {
-            "model": "openai/o3-pro",
-            # o3-pro for comprehensive analysis and superior drafting
-            "thinking_effort": "high",  # Universal parameter, translates to reasoning object
-            "verbosity": "high",  # Detailed comprehensive briefs
-            "disable_tools": True,  # o3-pro doesn't support tool calling
-        },
-        # Caseplan - LLM-driven workflow planning
-        # UPGRADED: Oct 2025 - Sonnet 4 -> Sonnet 4.5 for better reasoning
-        "caseplan": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-sonnet-4",
-            "temperature": 0.5,
-            "enforce_citations": False,
-            "disable_tools": True,  # Check if Sonnet 4.5 still has tool calling issues
-        },
-        # Caseplan assessment - budget recommendation (Sonnet)
-        # UPGRADED: Oct 2025 - Sonnet 4 -> Sonnet 4.5 for better reasoning
-        "caseplan-assessment": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-sonnet-4",
-            "temperature": 0.5,
-            "enforce_citations": False,
-            "disable_tools": True,  # Check if Sonnet 4.5 still has tool calling issues
-        },
-        # Chain of Verification - fast, efficient question generation
-        # UPGRADED: Oct 2025 - Sonnet 4 -> Sonnet 4.5 for better reasoning
-        "cove": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-sonnet-4",
-            "temperature": 0.2,
-            "top_p": 0.8,
-            "thinking_effort": "medium",  # General CoVe coordination
-            "enforce_citations": False,  # Avoid recursive verification
-            "disable_tools": True,  # Check if Sonnet 4.5 still has tool calling issues
-        },
-        # CoVe sub-stages with separate model control
-        # UPGRADED: Oct 2025 - Sonnet 4 -> Sonnet 4.5 for better question generation
-        "cove-questions": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-sonnet-4",
-            "temperature": 0.2,
-            "top_p": 0.8,
-            "thinking_effort": "low",  # Fast question generation, minimal thinking needed
-            "enforce_citations": False,
-            "disable_tools": True,  # Check if Sonnet 4.5 still has tool calling issues
-        },
-        # KEPT: Oct 2025 - GPT-5 base is optimal for fast, accurate answers (1.4% hallucination)
-        "cove-answers": {
-            "model": "openai/gpt-5",  # GPT-5 base for <1.4% hallucination rate
-            "temperature": 0.2,
-            "top_p": 0.3,
-            "thinking_effort": "high",  # Not max - save Pro for final verification
-            "enforce_citations": False,
-        },
-        # UPGRADED: Oct 2025 - Sonnet 4 -> Sonnet 4.5 for better inconsistency detection
-        "cove-verify": {
-            "model": "anthropic/claude-sonnet-4.5",
-            # OLD: "model": "anthropic/claude-sonnet-4",
-            "temperature": 0.2,
-            "top_p": 0.3,
-            "thinking_effort": "high",  # Critical inconsistency detection needs careful analysis
-            "enforce_citations": False,
-            "disable_tools": True,  # Check if Sonnet 4.5 still has tool calling issues
-        },
-        # UPGRADED: Oct 2025 - GPT-5 -> GPT-5 Pro for critical final output (<1% hallucination)
-        "cove-final": {
-            "model": "openai/gpt-5-pro",
-            # OLD: "model": "openai/gpt-5",
-            "temperature": 0.2,
-            "top_p": 0.4,
-            "thinking_effort": "max",  # Mandatory deep reasoning for final verification
-            "enforce_citations": False,
-        },
-    }
-
-    @classmethod
-    def for_command(
-        cls, command_name: str, sub_type: str = None, **overrides
-    ) -> "LLMClient":
-        """
-        Create an LLMClient configured for a specific command.
-
-        Args:
-            command_name: The command name (e.g., 'extractfacts', 'strategy')
-            sub_type: Optional sub-type for commands with multiple clients
-                     (e.g., 'orthodox', 'unorthodox', 'analysis' for brainstorm,
-                      'summary', 'issues' for digest)
-            **overrides: Any parameter overrides to apply to the default configuration
-
-        Returns:
-            Configured LLMClient instance with command context set
-
-        Examples:
-            # Simple command
-            client = LLMClientFactory.for_command('extractfacts')
-
-            # Command with sub-type
-            client = LLMClientFactory.for_command('brainstorm', 'orthodox')
-
-            # With overrides
-            client = LLMClientFactory.for_command('draft', temperature=0.7)
-        """
-        # Build the configuration key
-        if sub_type:
-            config_key = f"{command_name}-{sub_type}"
-        else:
-            config_key = command_name
-
-        # Get the configuration or fall back to a default
-        if config_key not in cls.COMMAND_CONFIGS:
-            # Default configuration for unknown commands
-            # UPGRADED: Oct 2025 - Default to Sonnet 4.5
-            config = {
-                "model": "anthropic/claude-sonnet-4.5",
-                # OLD: "model": "anthropic/claude-sonnet-4",
-                "temperature": 0.3,
-                "top_p": 0.7,
-                "disable_tools": True,  # Check if Sonnet 4.5 still has tool calling issues
-            }
-            # Use default configuration for commands without specific config
-            # This is expected behavior for many commands
-        else:
-            config = cls.COMMAND_CONFIGS[config_key].copy()
-
-        # Extract special flags
-        enforce_citations = config.pop("enforce_citations", False)
-        disable_tools = config.pop("disable_tools", False)
-
-        # Remove premium_model key if present (no longer needed)
-        config.pop("premium_model", None)
-
-        # Allow environment variable overrides for model selection
-        env_model_key = f"LITASSIST_{command_name.upper()}_MODEL"
-        if sub_type:
-            env_model_key = f"LITASSIST_{command_name.upper()}_{sub_type.upper()}_MODEL"
-
-        env_model = os.environ.get(env_model_key)
-        if env_model:
-            config["model"] = env_model
-            # Suppress informational message during pytest runs
-            if not os.environ.get("PYTEST_CURRENT_TEST"):
-                logger.info(info_message(f"Using model from environment: {env_model}"))
-
-        # Apply any provided overrides
-        config.update(overrides)
-
-        # Extract model from config
-        model = config.pop("model")
-
-        # Create the LLM client with remaining config as parameters
-        client = LLMClient(model, **config)
-
-        # Set the command context
-        client.command_context = config_key
-
-        # Set citation enforcement flag - explicitly set both True and False
-        client._enforce_citations = enforce_citations
-        client._disable_tools = disable_tools
-
-        return client
-
-    @classmethod
-    def get_model_for_command(cls, command_name: str, sub_type: str = None) -> str:
-        """
-        Get the model name configured for a specific command.
-
-        Useful for logging or debugging purposes.
-
-        Args:
-            command_name: The command name
-            sub_type: Optional sub-type
-
-        Returns:
-            Model name string
-        """
-        config_key = f"{command_name}-{sub_type}" if sub_type else command_name
-        config = cls.COMMAND_CONFIGS.get(
-            config_key, {"model": "anthropic/claude-sonnet-4.5", "disable_tools": True}
-        )
-        return config["model"]
-
-    @classmethod
-    def list_configurations(cls) -> Dict[str, Dict[str, Any]]:
-        """
-        List all available command configurations.
-
-        Returns:
-            Dictionary of all command configurations
-        """
-        return cls.COMMAND_CONFIGS.copy()
-
+# Parameter handling functions now imported from parameter_handler.py
+# Model configuration data imported from model_profiles.py
 
 class LLMClient(LLMVerificationMixin):
     """
